DEV: refactor to use new layout, and button replace
