+++ common/common.scss
@@ -22,3 +22,60 @@
   color: $secondary;
   font-size: $font-down-1;
 }
+
+.discourse-local-date {
+  > * {
+    pointer-events: none;
+  }
+
+  &.cooked-date {
+    color: var(--primary);
+    cursor: pointer;
+    border-bottom: 1px dashed var(--primary-medium);
+    white-space: nowrap;
+
+    .d-icon {
+      color: var(--primary);
+    }
+
+    &.past {
+      border-bottom-color: var(--primary-low-mid);
+    }
+
+    &.past[data-countdown] {
+      color: var(--primary-medium);
+    }
+  }
+}
+
+.locale-dates-previews {
+  max-width: 250px;
+
+  .preview {
+    display: flex;
+    flex-direction: column;
+    padding: 5px;
+    margin: 0;
+
+    .timezone {
+      font-weight: 700;
+    }
+
+    &.current {
+      background: var(--tertiary-low);
+    }
+  }
+}
+
+.download-calendar {
+  text-align: right;
+  cursor: pointer;
+  margin-top: 0.5em;
+}
+
+// @if (#{$replace_date_time}) {
+//   .btn.local-dates {
+//     display: none;
+//   }
+// }
+
