--- javascripts/discourse/components/modal/jitsi-event.hbs
@@ -1,63 +1,0 @@
-<DModal
-  @title={{theme-i18n "event_modal.title"}}
-  class="post-event-builder-modal"
-  @closeModal={{@closeModal}}
-  @flash={{this.flash}}
->
-  <:body>
-    <ConditionalLoadingSection @isLoading={{@model.event.isSaving}}>
-      <form>
-        <DateTimeInputRange
-          @from={{this.startsAt}}
-          @to={{this.endsAt}}
-          @toTimeFirst={{true}}
-          @clearable={{true}}
-          @timezone={{@model.event.timezone}}
-          @onChange={{this.onChangeDates}}
-        />
-
-        <EventField class="room-id" @label={{theme-prefix "room_label"}}>
-          <Input @value={{this.jitsiRoom}} />
-          <div class="desc">{{theme-i18n "modal.room_field_description"}}</div>
-        </EventField>
-
-        <EventField
-          class="recurrence"
-          @label="discourse_calendar.discourse_post_event.builder_modal.recurrence.label"
-        >
-          <ComboBox
-            class="available-recurrences"
-            @value={{@model.event.recurrence}}
-            @content={{this.availableRecurrences}}
-            @options={{hash
-              none="discourse_calendar.discourse_post_event.builder_modal.recurrence.none"
-            }}
-          />
-        </EventField>
-
-        <EventField
-          class="minimal-event"
-          @label="discourse_calendar.discourse_post_event.builder_modal.minimal.label"
-        >
-          <label class="checkbox-label">
-            <Input @type="checkbox" @checked={{@model.event.minimal}} />
-            <span class="message">
-              {{i18n
-                "discourse_calendar.discourse_post_event.builder_modal.minimal.checkbox_label"
-              }}
-            </span>
-          </label>
-        </EventField>
-      </form>
-    </ConditionalLoadingSection>
-  </:body>
-  <:footer>
-    <DButton
-      @type="button"
-      class="btn-primary"
-      @label="discourse_calendar.discourse_post_event.builder_modal.create"
-      @icon="calendar-day"
-      @action={{this.createEvent}}
-    />
-  </:footer>
-</DModal>
