+++ settings.yml
@@ -5,18 +5,15 @@ locations:
   default: "Header Icon|Composer Toolbar|Composer Gear Menu|Chat Icon|Chat Toolbar"
 only_available_to_staff:
   default: false
-discourse_post_event_enabled:
-  description: "Enable to allow creating meeting events through discourse-calendar integration"
+replace_date_time:
   default: false
+  description: 'Replace the "Insert date/time" button with a version that allows creating jitsi events'
 meet_jitsi_domain:
   default: "meet.jit.si"
   description: "Domain only (no protocol, no trailing slash)."
 jitsi_script_src:
   default: "https://meet.jit.si/external_api.js"
   description: "URL of external API javascript file. If you change this, you will also need to add the URL to the 'content security script src' site setting."
-button_icon:
-  default: "video"
-  description: "Enter the name of the FontAwesome 5 icon to display in the Jitsi button. "
 hide_iframe_buttons:
   default: false
   description: "Hide the choice for iframe settings from users, the default settings below will be used"
@@ -26,6 +23,12 @@ default_mobile_iframe:
 default_desktop_iframe:
   default: true
   description: "Enable iframe mode by default for desktop"
+button_icon:
+  default: "video"
+  description: "Enter the name of the FontAwesome 5 icon to display in the Jitsi button. "
+event_button_icon:
+  default: "calendar-alt"
+  description: "Enter the name of the FontAwesome 5 icon to display in the Jitsi event button. "
 copy_button_icon:
   default: "copy"
   description: "Enter the name of the FontAwesome 5 icon to display as the copy button. "
@@ -33,4 +36,4 @@ svg_icons:
   default: "video"
   type: "list"
   list_type: "compact"
-  description: "If using a custom icon, add it to this field too (in addition to 'button_icon' above."
+  description: "If using a custom icon, add it to this field too (in addition to 'button_icon', 'copy_button_icon', and 'event_button_icon' above.)"
